test(cli): empty-schema -> incomplete-schemas
